# protocol/instance/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ubichill Instance Runtime",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3002"
    }
  ],
  "security": [
    {
      "adminToken": []
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "health",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/realtime/v1/instances": {
      "get": {
        "operationId": "listInstances",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "instances": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RuntimePresence"
                      }
                    }
                  },
                  "required": [
                    "instances"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/realtime/v1/instances/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "put": {
        "operationId": "provisionInstance",
        "description": "Idempotent create. Existing instance state is never reset. Management token required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceDefinition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimePresence"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getInstance",
        "description": "Management token required in managed mode. Guest mode returns metadata with memberIds redacted.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimePresence"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "closeInstance",
        "parameters": [
          {
            "name": "emptyBefore",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Atomically reject if occupied or emptied after this epoch millisecond timestamp."
          }
        ],
        "responses": {
          "204": {
            "description": "Closed or already absent"
          },
          "409": {
            "description": "Occupied or recently emptied"
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/realtime/v1/instances/{id}/tickets": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "operationId": "issueTicket",
        "description": "Management API. Caller has already authorized the SNS user. Ticket expires in 2 minutes and is checked at admission. Existing connections remain admitted.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceGrant"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/realtime/v1/instances/{id}/guest": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "operationId": "joinGuest",
        "security": [],
        "description": "Only when started with --guests. Server issues an identity scoped to this room. Resume token expires in 24 hours and is renewed here.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceGrant"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "adminToken": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "RuntimePresence": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "x-go-name": "ID"
          },
          "name": {
            "type": "string"
          },
          "maxUsers": {
            "type": "integer"
          },
          "memberIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "x-go-name": "MemberIDs"
          },
          "emptySince": {
            "type": "integer",
            "format": "int64",
            "description": "Epoch milliseconds; 0 while occupied, includes reconnect grace."
          }
        },
        "required": [
          "id",
          "name",
          "maxUsers",
          "memberIds",
          "emptySince"
        ],
        "additionalProperties": false
      },
      "InstanceGrant": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "expiresAt": {
            "type": "integer",
            "format": "int64"
          },
          "protocolVersion": {
            "type": "integer",
            "enum": [
              1
            ]
          }
        },
        "required": [
          "url",
          "token",
          "userId",
          "expiresAt",
          "protocolVersion"
        ],
        "additionalProperties": false
      },
      "Snapshot": {
        "type": "object",
        "properties": {
          "entities": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "availableComponents": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "activeMods": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "environment": {
            "type": "object",
            "additionalProperties": true
          },
          "lock": {
            "type": "object",
            "additionalProperties": true,
            "x-go-type": "json.RawMessage",
            "x-go-type-skip-optional-pointer": true
          },
          "sourceKind": {
            "type": "string",
            "x-go-type-skip-optional-pointer": true
          }
        },
        "required": [
          "entities",
          "availableComponents",
          "activeMods",
          "environment"
        ],
        "additionalProperties": false
      },
      "InstanceDefinition": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "x-go-name": "ID"
          },
          "name": {
            "type": "string"
          },
          "maxUsers": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000
          },
          "snapshot": {
            "$ref": "#/components/schemas/Snapshot"
          }
        },
        "required": [
          "id",
          "name",
          "maxUsers",
          "snapshot"
        ],
        "additionalProperties": false
      },
      "GuestRequest": {
        "type": "object",
        "properties": {
          "resumeToken": {
            "type": "string"
          }
        },
        "required": [],
        "additionalProperties": false
      },
      "TicketRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "userId"
        ],
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "error"
        ],
        "additionalProperties": false
      }
    }
  }
}