# services/instance/internal/protocol/api.gen.go
// Package protocol provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package protocol

import (
	"encoding/json"
)

const (
	AdminTokenScopes = "adminToken.Scopes"
)

// Defines values for ErrorSuccess.
const (
	False ErrorSuccess = false
)

// Defines values for InstanceGrantProtocolVersion.
const (
	N1 InstanceGrantProtocolVersion = 1
)

// Error defines model for Error.
type Error struct {
	Error   string       `json:"error"`
	Success ErrorSuccess `json:"success"`
}

// ErrorSuccess defines model for Error.Success.
type ErrorSuccess bool

// GuestRequest defines model for GuestRequest.
type GuestRequest struct {
	ResumeToken *string `json:"resumeToken,omitempty"`
}

// InstanceDefinition defines model for InstanceDefinition.
type InstanceDefinition struct {
	ID       string   `json:"id"`
	MaxUsers int      `json:"maxUsers"`
	Name     string   `json:"name"`
	Snapshot Snapshot `json:"snapshot"`
}

// InstanceGrant defines model for InstanceGrant.
type InstanceGrant struct {
	ExpiresAt       int64                        `json:"expiresAt"`
	ProtocolVersion InstanceGrantProtocolVersion `json:"protocolVersion"`
	Token           string                       `json:"token"`
	Url             string                       `json:"url"`
	UserId          string                       `json:"userId"`
}

// InstanceGrantProtocolVersion defines model for InstanceGrant.ProtocolVersion.
type InstanceGrantProtocolVersion int

// RuntimePresence defines model for RuntimePresence.
type RuntimePresence struct {
	// EmptySince Epoch milliseconds; 0 while occupied, includes reconnect grace.
	EmptySince int64    `json:"emptySince"`
	ID         string   `json:"id"`
	MaxUsers   int      `json:"maxUsers"`
	MemberIDs  []string `json:"memberIds"`
	Name       string   `json:"name"`
}

// Snapshot defines model for Snapshot.
type Snapshot struct {
	ActiveMods          []string                 `json:"activeMods"`
	AvailableComponents []map[string]interface{} `json:"availableComponents"`
	Entities            []map[string]interface{} `json:"entities"`
	Environment         map[string]interface{}   `json:"environment"`
	Lock                json.RawMessage          `json:"lock,omitempty"`
	SourceKind          string                   `json:"sourceKind,omitempty"`
}

// TicketRequest defines model for TicketRequest.
type TicketRequest struct {
	UserId string `json:"userId"`
}

// CloseInstanceParams defines parameters for CloseInstance.
type CloseInstanceParams struct {
	// EmptyBefore Atomically reject if occupied or emptied after this epoch millisecond timestamp.
	EmptyBefore *int64 `form:"emptyBefore,omitempty" json:"emptyBefore,omitempty"`
}

// ProvisionInstanceJSONRequestBody defines body for ProvisionInstance for application/json ContentType.
type ProvisionInstanceJSONRequestBody = InstanceDefinition

// JoinGuestJSONRequestBody defines body for JoinGuest for application/json ContentType.
type JoinGuestJSONRequestBody = GuestRequest

// IssueTicketJSONRequestBody defines body for IssueTicket for application/json ContentType.
type IssueTicketJSONRequestBody = TicketRequest